sciconsist_pilot/pyproject.toml
[project]
name = "sciconsist-pilot"
version = "0.1.0"
description = "SciConsist Pilot: Cross-Model Factual Entailment Head for Scientific Document Verification"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.45.0",
    "datasets>=3.0.0",
    "accelerate>=1.0.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "scikit-learn>=1.5.0",
    "scipy>=1.14.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "tqdm>=4.66.0",
    "rich>=13.0.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
